tweaked Wagn::Conf / default environment
